llm_port_mcp/pyproject.toml
[project]
name = "llm_port_mcp"
version = "0.1.0"
description = "MCP Tool Registry — governed MCP broker for llm.port"
readme = "README.md"
requires-python = ">=3.11,<4"
authors = [
  { name = "Sachith Liyanagama", email = "liyanagama@outlook.de" },
]
maintainers = [
  { name = "Sachith Liyanagama", email = "liyanagama@outlook.de" },
]
dependencies = [
  "fastapi>=0.115.6",
  "uvicorn[standard]>=0.34.0",
  "pydantic>=2.10.4",
  "pydantic-settings>=2.7.0",
  "yarl>=1.18.3",
  "ujson>=5.10.0",
  "prometheus-client>=0.21.1",
  "prometheus-fastapi-instrumentator==7.0.0",
  "sentry-sdk>=2.19.2",
  "opentelemetry-api>=1.29.0",
  "opentelemetry-sdk>=1.29.0",
  "opentelemetry-exporter-otlp>=1.29.0",
  "opentelemetry-instrumentation>=0.50b0",
  "opentelemetry-instrumentation-fastapi>=0.50b0",
  "loguru>=0.7.3",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.30.0",
  "alembic>=1.14.0",
  "cryptography>=44.0.0",
  "httpx>=0.27.0",
  "pyjwt>=2.9.0",
  "redis[hiredis]>=5.0.0",
  "mcp>=1.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8",
  "ruff>=0.5.0",
  "mypy>=1.10.1",
  "pre-commit>=3.7.1",
  "black>=24.4.2",
  "pytest-cov>=5",
  "anyio>=4",
  "pytest-env>=1.1.3",
  "httpx>=0.27.0",
  "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["llm_port_mcp",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "LLM_PORT_MCP_ENVIRONMENT=pytest",
    "LLM_PORT_MCP_DB_BASE=llm_mcp_test",
    "LLM_PORT_MCP_SENTRY_DSN=",
]

[tool.ruff]
lint.select = [
    "E",   # Error
    "F",   # Pyflakes
    "W",   # Pycodestyle
    "C90", # McCabe complexity
    "I",   # Isort
    "N",   # pep8-naming
    "D",   # Pydocstyle
    "ANN", # Pytype annotations
    "S",   # Bandit
    "B",   # Bugbear
    "COM", # Commas
    "C4",  # Comprehensions
    "ISC", # Implicit string concat
    "PIE", # Unnecessary code
    "T20", # Catch prints
    "PYI", # validate pyi files
    "Q",   # Checks for quotes
    "RSE", # Checks raise statements
    "RET", # Checks return statements
    "SLF", # Self checks
    "SIM", # Simplificator
    "PTH", # Pathlib checks
    "ERA", # Checks for commented out code
    "PL",  # PyLint checks
    "RUF", # Specific to Ruff checks
]
lint.ignore = [
    "D105",
    "D107",
    "B008",
    "D211",
    "D212",
    "D401",
    "D104",
    "D100",
    "D202",
    "ANN102",
    "ANN101",
    "ANN401",
    "PLR0913",
    "D106",
]
exclude = [
    ".venv/"
]
lint.mccabe = { max-complexity = 10 }
line-length = 88

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "S101",
]

[tool.ruff.lint.pydocstyle]
convention = "pep257"
ignore-decorators = ["typing.overload"]

[tool.ruff.lint.pylint]
allow-magic-value-types = ["int", "str", "float", "bytes"]
